// src/main/web-app/src/component/github/models/repository-item-model.ts
export interface RepositoryItemModel {
    name: string;
    full_name: string;
    stargazers_count: number;
    stargazers_url: string;
    score: number;
    private: boolean;
    language: string;
    homepage: string;
    git_url: string;
    html_url: string;
    url: string;
    description: string;
    default_branch: string;
    created_at: string;
    forks_count: number;
    forks_url:string
}